+++ spanner-core/src/test/java/com/google/cloud/spark/spanner/acceptance/DataprocAcceptanceTestBase.java
@@ -282,7 +282,7 @@ public class DataprocAcceptanceTestBase {
       String connectorJarUri)
       throws Exception {
     String clusterName = generateClusterName(testId);
-    System.out.printf(
+    logger.info(
         "createClusterIfNeeded clusterName: %s, testId: %s, dataprocImageVersion: %s, connectorJarUri: %s, PROJECT_ID: %s, REGION: %s%n",
         clusterName, testId, dataprocImageVersion, connectorJarUri, PROJECT_ID, REGION);
     cluster(
