+++ spanner-spark-common/pom.xml
@@ -47,7 +47,6 @@
     <dependency>
       <groupId>com.google.code.gson</groupId>
       <artifactId>gson</artifactId>
-      <version>2.10.1</version>
     </dependency>
 
     <dependency>
