+++ spanner-spark-common/src/main/java/com/google/cloud/spark/spanner/graph/SpannerGraphScanner.java
@@ -4,9 +4,22 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.api.gax.grpc.GrpcCallContext;
 import com.google.api.gax.rpc.ApiCallContext;
 import com.google.cloud.Tuple;
-import com.google.cloud.spanner.*;
+import com.google.cloud.spanner.BatchReadOnlyTransaction;
+import com.google.cloud.spanner.BatchTransactionId;
+import com.google.cloud.spanner.Options;
+import com.google.cloud.spanner.Partition;
+import com.google.cloud.spanner.PartitionOptions;
+import com.google.cloud.spanner.SpannerOptions;
 import com.google.cloud.spanner.SpannerOptions.CallContextConfigurator;
-import com.google.cloud.spark.spanner.*;
+import com.google.cloud.spanner.Statement;
+import com.google.cloud.spanner.TimestampBound;
+import com.google.cloud.spark.spanner.BatchClientWithCloser;
+import com.google.cloud.spark.spanner.SpannerInputPartitionContext;
+import com.google.cloud.spark.spanner.SpannerPartition;
+import com.google.cloud.spark.spanner.SpannerPartitionReaderFactory;
+import com.google.cloud.spark.spanner.SpannerRowConverter;
+import com.google.cloud.spark.spanner.SpannerScanner;
+import com.google.cloud.spark.spanner.SpannerUtils;
 import com.google.cloud.spark.spanner.graph.query.SpannerGraphQuery;
 import com.google.common.collect.ImmutableSet;
 import io.grpc.Context;
