+++ spanner-spark-common/src/main/java/com/google/cloud/spark/spanner/graph/query/SpannerGraphQuery.java
@@ -10,7 +10,13 @@ import com.google.cloud.spark.spanner.graph.SpannerGraphConfigs;
 import com.google.cloud.spark.spanner.graph.SpannerGraphConfigs.LabelConfig;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Set;
 import java.util.stream.Collectors;
 import org.apache.spark.sql.types.IntegralType;
 import org.apache.spark.sql.types.StructField;
