+++ spark-3.1-spanner-lib/pom.xml
@@ -45,7 +45,6 @@
     <dependency>
       <groupId>com.google.code.gson</groupId>
       <artifactId>gson</artifactId>
-      <version>2.10.1</version>
     </dependency>
 
     <dependency>
