+++ spark-3.1-spanner-lib/src/main/java/com/google/cloud/spark/spanner/Spark31SpannerTableProvider.java
@@ -14,82 +14,4 @@
 
 package com.google.cloud.spark.spanner;
 
-import com.google.cloud.spark.spanner.graph.SpannerGraphBuilder;
-import java.util.Map;
-import org.apache.spark.sql.connector.catalog.Table;
-import org.apache.spark.sql.connector.catalog.TableProvider;
-import org.apache.spark.sql.connector.expressions.Transform;
-import org.apache.spark.sql.sources.DataSourceRegister;
-import org.apache.spark.sql.types.StructType;
-import org.apache.spark.sql.util.CaseInsensitiveStringMap;
-
-public class Spark31SpannerTableProvider implements DataSourceRegister, TableProvider {
-
-  /*
-   * Infers the schema of the table identified by the given options.
-   */
-  @Override
-  public StructType inferSchema(CaseInsensitiveStringMap options) {
-    return getTable(options).schema();
-  }
-
-  /*
-   * Returns a Table instance with the specified table schema,
-   * partitioning and properties to perform a read or write.
-   */
-  @Override
-  public Table getTable(
-      StructType schema, Transform[] partitioning, Map<String, String> properties) {
-    CaseInsensitiveStringMap options = new CaseInsensitiveStringMap(properties);
-    boolean enablePartialRowUpdates =
-        Boolean.parseBoolean(options.getOrDefault("enablePartialRowUpdates", "false"));
-
-    boolean hasTable = options.containsKey("table");
-    boolean hasGraph = options.containsKey("graph");
-    if (hasTable && !hasGraph) {
-      if (enablePartialRowUpdates) {
-        return new SpannerTable(options, schema);
-      } else {
-        return new SpannerTable(options);
-      }
-    } else if (!hasTable && hasGraph) {
-      return SpannerGraphBuilder.build(options);
-    } else {
-      throw new SpannerConnectorException(
-          SpannerErrorCode.INVALID_ARGUMENT,
-          "properties must contain one of \"table\" or \"graph\"");
-    }
-  }
-
-  /*
-   * Returns true if the source has the ability of
-   * accepting external table metadata when getting tables.
-   */
-  @Override
-  public boolean supportsExternalMetadata() {
-    return true;
-  }
-
-  /*
-   * Implements DataSourceRegister.shortName(). This method allows Spark to match
-   * the DataSource when spark.read(...).format("spanner") is invoked.
-   */
-  @Override
-  public String shortName() {
-    return "cloud-spanner";
-  }
-
-  private Table getTable(Map<String, String> properties) {
-    boolean hasTable = properties.containsKey("table");
-    boolean hasGraph = properties.containsKey("graph");
-    if (hasTable && !hasGraph) {
-      return new SpannerTable(properties);
-    } else if (!hasTable && hasGraph) {
-      return SpannerGraphBuilder.build(properties);
-    } else {
-      throw new SpannerConnectorException(
-          SpannerErrorCode.INVALID_ARGUMENT,
-          "properties must contain one of \"table\" or \"graph\"");
-    }
-  }
-}
+public class Spark31SpannerTableProvider extends SpannerTableProviderBase {}
