+++ spark-3.2-spanner-lib/pom.xml
@@ -42,7 +42,6 @@
     <dependency>
       <groupId>com.google.code.gson</groupId>
       <artifactId>gson</artifactId>
-      <version>2.10.1</version>
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
