+++ spark-3.2-spanner/pom.xml
@@ -36,153 +36,6 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-shade-plugin</artifactId>
-        <configuration>
-          <skip>${shade.skip}</skip>
-          <shadedArtifactAttached>false</shadedArtifactAttached>
-          <relocations>
-            <relocation>
-              <pattern>com.google.common</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.com.google.common</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>android</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.android</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>autovalue</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.autovalue</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>avro.shaded</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.avro.shaded</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>com.fasterxml</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.com.fasterxml</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>com.github</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.com.github</shadedPattern>
-              <!-- Need to exclude the zstd library from being repackaged because of https://github.com/luben/zstd-jni#limitations -->
-              <excludes>
-                <exclude>com.github.luben.zstd.**</exclude>
-              </excludes>
-            </relocation>
-            <relocation>
-              <pattern>com.google</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.com.google</shadedPattern>
-              <excludes>
-                <exclude>com.google.cloud.bigquery.connector.common.**</exclude>
-                <exclude>com.google.cloud.spark.spanner.**</exclude>
-              </excludes>
-            </relocation>
-            <relocation>
-              <pattern>com.thoughtworks.paranamer</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.com.thoughtworks.paranamer
-              </shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>com.typesafe</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.com.typesafe</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>io.grpc</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.io.grpc</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>io.netty</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.io.netty</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>io.opencensus</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.io.opencensus</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>io.perfmark</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.io.perfmark</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.apache.arrow</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.apache.arrow</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.apache.beam</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.apache.beam</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.apache.commons</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.apache.commons</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.apache.http</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.apache.http</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.checkerframework</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.checkerframework
-              </shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.codehaus.mojo</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.codehaus.mojo</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.json</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.json</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.threeten</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.threeten</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.tukaani.xz</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.tukaani.xz</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>org.xerial.snappy</pattern>
-              <shadedPattern>com.google.cloud.spark.spanner.repackaged.org.xerial.snappy</shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>META-INF/native/libio_grpc_netty_shaded_netty</pattern>
-              <shadedPattern>
-                META-INF/native/libcom_google_cloud_spark_spanner_repackaged_io_grpc_netty_shaded_io_grpc_netty
-              </shadedPattern>
-            </relocation>
-            <relocation>
-              <pattern>META-INF/native/io_grpc_netty_shaded_netty</pattern>
-              <shadedPattern>
-                META-INF/native/com_google_cloud_spark_spanner_repackaged_io_grpc_netty_shaded_io_grpc_netty
-              </shadedPattern>
-            </relocation>
-          </relocations>
-          <transformers>
-            <transformer
-              implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
-          </transformers>
-          <filters>
-            <filter>
-              <artifact>*:*</artifact>
-              <artifact>*:*</artifact>
-              <excludes>
-                <exclude>module-info.class</exclude>
-                <exclude>META-INF/*.MF</exclude>
-                <exclude>META-INF/*.SF</exclude>
-                <exclude>META-INF/*.DSA</exclude>
-                <exclude>META-INF/*.RSA</exclude>
-                <exclude>META-INF/maven/**</exclude>
-                <exclude>**/*.proto</exclude>
-              </excludes>
-            </filter>
-          </filters>
-        </configuration>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-          </execution>
-        </executions>
       </plugin>
     </plugins>
   </build>
