+++ spark-3.5-spanner-lib/pom.xml
@@ -61,7 +61,6 @@
     <dependency>
       <groupId>com.google.code.gson</groupId>
       <artifactId>gson</artifactId>
-      <version>2.10.1</version>
     </dependency>
   </dependencies>
 </project>
